feat(crm): add PDF download button to opportunity quotation tab
Add the ability to download the quotation PDF from the opportunity detail
view (credit tab). Refactored generateAmortizationTable to a shared module
to avoid code duplication between quoter and CreditDetailView.
