+++ apps/crm/apps/web/src/lib/generate-pdf.ts
@@ -1,6 +1,65 @@
 import jsPDF from "jspdf";
 import autoTable from "jspdf-autotable";
 
+// Tipo para filas de la tabla de amortización
+export interface AmortizationRow {
+	period: number;
+	initialBalance: number;
+	interestPlusVAT: number;
+	principal: number;
+	finalBalance: number;
+}
+
+// Función para generar tabla de amortización
+export function generateAmortizationTable(
+	totalFinanced: number,
+	monthlyRate: number,
+	termMonths: number,
+): AmortizationRow[] {
+	const table: AmortizationRow[] = [];
+	let balance = totalFinanced;
+	const r = monthlyRate / 100;
+	const VAT = 0.12; // 12% IVA
+
+	// Calcular la cuota base (sin seguro ni GPS)
+	const rWithVAT = r * (1 + VAT);
+	const factor = (1 + rWithVAT) ** termMonths;
+	const baseMonthlyPayment =
+		(totalFinanced * (rWithVAT * factor)) / (factor - 1);
+
+	// Período 0 (inicial)
+	const initialInterest = balance * r;
+	const initialInterestWithVAT = initialInterest * (1 + VAT);
+
+	table.push({
+		period: 0,
+		initialBalance: Math.round(balance * 100) / 100,
+		interestPlusVAT: Math.round(initialInterestWithVAT * 100) / 100,
+		principal: 0,
+		finalBalance: Math.round(balance * 100) / 100,
+	});
+
+	// Períodos 1 a termMonths
+	for (let i = 1; i <= termMonths; i++) {
+		const interest = balance * r;
+		const interestWithVAT = interest * (1 + VAT);
+		const principalPayment = baseMonthlyPayment - interestWithVAT;
+		const newBalance = balance - principalPayment;
+
+		table.push({
+			period: i,
+			initialBalance: Math.round(balance * 100) / 100,
+			interestPlusVAT: Math.round(interestWithVAT * 100) / 100,
+			principal: Math.round(principalPayment * 100) / 100,
+			finalBalance: Math.round((newBalance > 0 ? newBalance : 0) * 100) / 100,
+		});
+
+		balance = newBalance;
+	}
+
+	return table;
+}
+
 // Define an interface for the quotation data
 interface QuotationData {
 	vehicleBrand: string | null;
